Supports Readline installed via MacPorts
MacPorts install readline under `/opt/local/lib` so I added that
path to the list of paths that are searched for the readline shared
library.

I also refactored the code a bit to just to keep the logic consistent
and allow Raku to find readline in the default library path.