Support basic use with Nix on macOS
Previously, any use of Readline on macOS would only use the GNU
readline library installed by Homebrew.

This is an initial patch to detect use in a Nix environment and look
for the library via `LD_LIBRARY_PATH` instead in that case.  Further
work is probably required for ideal Nix support.